Remove public groups, split main into smaller parts.
The public groups feature will be split off into a standalone bot.
